Clean up #ifdef mess in RegistersView
