+++ src/cross/widgets/Configuration.cpp
@@ -143,6 +143,9 @@ Configuration::Configuration(const ConfigurationPalette & p) : QObject(), noMore
     defaultColors.insert("RegistersHighlightWriteColor", QColor("#B00000"));
     defaultColors.insert("RegistersHighlightReadWriteColor", QColor("#808000"));
 
+    defaultColors.insert("TraceNewValueColor", QColor("#FF0000"));
+    defaultColors.insert("TraceNewValueBackgroundColor", Qt::transparent);
+
     defaultColors.insert("InstructionHighlightColor", white);
     defaultColors.insert("InstructionHighlightBackgroundColor", QColor("#CC0000"));
     defaultColors.insert("InstructionCommaColor", black);
